Add Playwright tests workflow and update baseURL
A new Github Action workflow that runs Playwright tests is implemented. This workflow triggers on push and pull requests to the 'gh-actions' and 'master' branches. In the Playwright configuration file, the base URL for testing is also being updated from an environment variable to a static URL - 'https://showcase.janus-idp.io/'.

Signed-off-by: Gustavo Lira <guga.java@gmail.com>